requirements.txt
requests>=2.31.0
pandas>=2.0.3
numpy>=1.24.3
python-dotenv>=1.0.0
ta>=0.10.2
schedule>=1.2.0